% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_theta2.R
\name{calculate_theta2}
\alias{calculate_theta2}
\title{Calculate optimal trait values (theta) given a fitted \code{coevfit} object and a vector of "intervention" trait values x-hat, which are assumed to be held constant.}
\usage{
calculate_theta2(object, x_hat)
}
\arguments{
\item{object}{An object of class \code{coevfit}}

\item{x_hat}{A numeric vector of intervention values where the elements correspond to the varibables of the coevfit. If an element is NA, then}
}
\value{
Posterior samples in matrix format
}
\description{
Calculate optimal trait values (theta) given a fitted \code{coevfit} object and a vector of "intervention" trait values x-hat, which are assumed to be held constant.
}
