% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coev_calculate_delta_theta.R
\name{coev_calculate_delta_theta}
\alias{coev_calculate_delta_theta}
\title{Calculate delta theta from a fitted \code{coevfit} object}
\usage{
coev_calculate_delta_theta(object, response, predictor)
}
\arguments{
\item{object}{An object of class \code{coevfit}}

\item{response}{A character string of length one equal to one of the
coevolving variables in the model (strictly case-sensitive)}

\item{predictor}{A character string of length one equal to one of the
coevolving variables in the model (strictly case-sensitive)}
}
\value{
Posterior samples in the draws_array format
}
\description{
Calculate \eqn{\Delta\theta} from a fitted \code{coevfit} object.
\eqn{\Delta\theta} is defined as the difference in the equilibrium value for
a "response" trait that results from a standardised increase in the value of
another "predictor" trait. This value can be used to assess contingencies
and directionality between variables in the coevolutionary process.
}
\details{
This function calculates \eqn{\Delta\theta}, which is defined as the
difference in the equilibrium value for a "response" trait that results
from a standardised increase in the value of another "predictor" trait. The
function first calculates the equilibrium trait value for the response
trait when the predictor trait is held at its empirical median value (the
\code{\link{coev_calculate_theta}} function is used for this purpose, see
\code{help(coev_calculate_theta)} for further details). The function then
calculates the equilibrium trait value for the response variable after
increasing the predictor trait from its median by one median absolute
deviation. The function then returns the posterior difference between these
values. The resulting \eqn{\Delta\theta} samples can be used to infer
whether increases in one trait have a positive or negative selective effect
on another trait in the model.
}
\examples{
\dontrun{
# fit dynamic coevolutionary model
fit <- coev_fit(
  data = authority$data,
  variables = list(
    political_authority = "ordered_logistic",
    religious_authority = "ordered_logistic"
  ),
  id = "language",
  tree = authority$phylogeny,
  # additional arguments for cmdstanr::sample()
  chains = 4,
  parallel_chains = 4,
  seed = 1
  )

# calculate delta theta
coev_calculate_delta_theta(
  object = fit,
  response = "political_authority",
  predictor = "religious_authority"
  )
}

}
\references{
Ringen, E., Martin, J. S., & Jaeggi, A. (2021). Novel phylogenetic methods
reveal that resource-use intensification drives the evolution of "complex"
societies. \emph{EcoEvoRXiv}. \code{doi:10.32942/osf.io/wfp95}

Sheehan, O., Watts, J., Gray, R. D., Bulbulia, J., Claessens, S., Ringen,
E. J., & Atkinson, Q. D. (2023). Coevolution of religious and political
authority in Austronesian societies. \emph{Nature Human Behaviour},
\emph{7}(1), 38-45. \code{10.1038/s41562-022-01471-y}
}
\seealso{
\code{\link{coev_calculate_delta_theta}},
\code{\link{coev_plot_delta_theta}}

Other calculate:
\code{\link[=coev_calculate_theta]{coev_calculate_theta()}}
}
\author{
Scott Claessens \email{scott.claessens@gmail.com}, Erik Ringen
\email{erikjacob.ringen@uzh.ch}
}
\concept{calculate}
