% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coev_identify_nodes.R
\name{coev_identify_nodes}
\alias{coev_identify_nodes}
\title{Interactively select multiple nodes from a plotted phylogeny}
\usage{
coev_identify_nodes(tree, n = NULL)
}
\arguments{
\item{tree}{A \code{phylo} object, typically the value returned (invisibly)
by \code{\link{coev_plot_node_labels}}.}

\item{n}{(optional) Maximum number of nodes to collect. If \code{NULL}
(default), the loop continues until the user cancels.}
}
\value{
An integer vector of unique node IDs, in the order they were
clicked.
}
\description{
A convenience wrapper around \code{\link[ape]{identify.phylo}} that
allows the user to click on \emph{multiple} internal nodes in succession
and returns the collected node IDs. By default,
\code{\link[ape]{identify.phylo}} returns after a single click; this
function calls it in a loop until the user signals completion (by
clicking outside the tree, pressing Esc, or right-clicking, depending
on graphics device) or until \code{n} clicks have been collected.
}
\details{
Intended workflow:

\preformatted{
tree <- coev_plot_node_labels(fit)
nodes <- coev_identify_nodes(tree)
coev_ancestral_states(fit, nodes = nodes)
}
}
\examples{
\dontrun{
# plot tree with internal node IDs, then click on nodes of interest
tree <- coev_plot_node_labels(fit)
nodes <- coev_identify_nodes(tree)
coev_ancestral_states(fit, nodes = nodes)
}

}
\seealso{
\code{\link{coev_plot_node_labels}},
\code{\link{coev_ancestral_states}},
\code{\link[ape]{identify.phylo}}
}
\author{
Erik Ringen \email{erikjacob.ringen@uzh.ch}
}
