% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coev_make_stancode.R
\name{coev_make_stancode}
\alias{coev_make_stancode}
\title{Make Stan code for dynamic coevolutionary model}
\usage{
coev_make_stancode(
  data,
  variables,
  id,
  tree,
  effects_mat = NULL,
  complete_cases = FALSE,
  lon_lat = NULL,
  dist_k = NA,
  dist_cov = "exp_quad",
  measurement_error = NULL,
  prior = NULL,
  scale = TRUE,
  estimate_correlated_drift = TRUE,
  estimate_residual = TRUE,
  log_lik = FALSE,
  prior_only = FALSE,
  dist_mat = deprecated()
)
}
\arguments{
\item{data}{An object of class \code{data.frame} (or one that can be coerced
to that class) containing data of all variables used in the model.}

\item{variables}{A named list identifying variables that should coevolve in
the model and their associated response distributions as character strings
(e.g. \code{list(var1 = "bernoulli_logit", var2 = "ordered_logistic")}).
Variable names must refer to valid column names in data. Currently, the
only supported response distributions are \code{bernoulli_logit},
\code{ordered_logistic}, \code{poisson_softplus},
\code{negative_binomial_softplus}, \code{normal}, and \code{gamma_log}.
Bernoulli variables must be 0/1 integers, ordered variables must be ordered
factors, Poisson and negative binomial variables must be positive integers,
normal variables must be continuous numeric, and gamma variables must be
positive numeric.}

\item{id}{A character of length one identifying the variable in the data that
links rows to tips on the phylogeny (strictly case-sensitive). Must refer
to a valid column name in the data. The id column must exactly match the
tip labels in the phylogeny.}

\item{tree}{A phylogenetic tree object of class \code{phylo} or
\code{multiPhylo}. The tree(s) must be rooted and must include positive
non-zero branch lengths. All trees in \code{multiPhylo} objects must have
the same number of internal nodes and branches.}

\item{effects_mat}{(optional) A boolean matrix with row and column names
exactly matching the variables declared for the model. If not specified,
all cross-lagged effects will be estimated in the model. If specified, the
model will only estimate cross-lagged effects where cells in the matrix are
TRUE and will ignore cross-lagged effects where cells in the matrix are
FALSE. In the matrix, columns represent predictor variables and rows
represent outcome variables. All autoregressive effects (e.g., X -> X) must
be TRUE in the matrix.}

\item{complete_cases}{(optional) Logical. If \code{FALSE} (default), all
missing values are imputed by the model. If \code{TRUE}, taxa with missing
data are excluded.}

\item{lon_lat}{(optional) A \code{data.frame} containing longitude and
latitude values for all tips on the phylogeny. The data frame must contain
the following columns: a column labelled "id" that exactly matches the tip
labels on the phylogeny; a column labelled "longitude" that declares
longitude values in decimal degrees; and a column labelled "latitude" that
declares latitude values in decimal degrees. If specified, the model will
additionally control for spatial location by including a separate Gaussian
Process over locations for every coevolving variable in the model.}

\item{dist_k}{An integer of length one specifying the number of basis
functions for computing Hilbert-space approximate Gaussian Processes over
locations. If \code{NA} (the default), exact Gaussian Processes are
computed.}

\item{dist_cov}{A string of length one specifying the covariance kernel used
for Gaussian Processes over locations (strictly case-sensitive). Currently
supported are \code{"exp_quad"} (exponentiated-quadratic kernel; default),
\code{"exponential"} (exponential kernel), and \code{"matern32"}
(Matern 3/2 kernel).}

\item{measurement_error}{(optional) A named list of coevolving variables and
their associated columns in the dataset containing standard errors. Only
valid for normally-distributed variables. For example, if we declare
\code{variables = list(x = "normal", y = "normal")}, then we could set
\code{measurement_error = list(x = "x_std_err")} to tell the function to
include measurement error on \code{x} using standard errors from the
\code{x_std_err} column of the dataset.}

\item{prior}{(optional) A named list of priors for the model. If not
specified, the model uses default priors (see \code{help(coev_fit)}).
Alternatively, the user can specify a named list of priors. The list must
contain non-duplicated entries for any of the following parameters: the
autoregressive effects (\code{A_diag}), the cross effects
(\code{A_offdiag}), the Cholesky factor for the drift matrix (\code{L_R}),
the drift std. dev. parameters (\code{Q_sigma}), the continuous time
intercepts (\code{b}), the ancestral states for the traits
(\code{eta_anc}), the cutpoints for ordinal variables (\code{c}), the
overdispersion parameters for negative binomial variables (\code{phi}),
the shape parameters for gamma variables (\code{shape}), the sigma
parameters for Gaussian Processes over locations (\code{sigma_dist}), the
rho parameters for Gaussian Processes over locations (\code{rho_dist}), the
residual standard deviations when there are repeated observations
(\code{sigma_residual}), and the Cholesky factor for the residual
correlations when there are repeated observations (\code{L_residual}).
These must be entered with valid prior strings, e.g.
\code{list(A_offdiag = "normal(0, 2)")}. Invalid prior strings will throw
an error when the function internally checks the syntax of resulting Stan
code.}

\item{scale}{Logical. If \code{TRUE} (default), variables following the
\code{normal} and \code{gamma_log} response distributions are scaled before
fitting the model. Continuous variables following the \code{normal}
distribution are standardised (e.g., mean centered and divided by their
standard deviation) and positive real variables following the
\code{gamma_log} distribution are divided by the mean value without
centering. This approach is recommended when using default priors to
improve efficiency and ensure accurate inferences. If \code{FALSE},
variables are left unscaled for model fitting. In this case, users should
take care to set sensible priors on variables.}

\item{estimate_correlated_drift}{Logical. If \code{TRUE} (default), the model
estimates the off-diagonals for the \deqn{Q} drift matrix (i.e., correlated
drift). If \code{FALSE}, the off-diagonals for the \deqn{Q} drift matrix
are set to zero.}

\item{estimate_residual}{Logical. If \code{TRUE} (default), the model
estimates residual standard deviations and residual correlations when there
are repeated observations for taxa. If \code{FALSE}, residual standard
deviations and residual correlations are not estimated. The latter may be
preferable in cases where repeated observations are sparse (e.g., only some
taxa have only few repeated observations). This argument only applies when
repeated observations are present in the data.}

\item{log_lik}{Logical. Set to \code{FALSE} by default. If \code{TRUE}, the
model returns the pointwise log likelihood, which can be used to calculate
WAIC and LOO.}

\item{prior_only}{Logical. If \code{FALSE} (default), the model is fitted to
the data and returns a posterior distribution. If \code{TRUE}, the model
samples from the prior only, ignoring the likelihood.}

\item{dist_mat}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use the \code{lon_lat}
argument instead.}
}
\value{
A character string containing the \pkg{Stan} code to fit the dynamic
coevolutionary model
}
\description{
Make the \pkg{Stan} code for the Bayesian dynamic coevolutionary model.
\pkg{Stan} code is generated, checked for syntactical errors, and then
returned as a character string.
}
\details{
For further details, see \code{help(coev_fit)}
}
\examples{
# make stan code
stan_code <- coev_make_stancode(
  data = authority$data,
  variables = list(
    political_authority = "ordered_logistic",
    religious_authority = "ordered_logistic"
  ),
  id = "language",
  tree = authority$phylogeny
)

# include effects matrix
effects_mat <-
  matrix(
    c(TRUE, TRUE,
      FALSE, TRUE),
    nrow = 2,
    dimnames = list(
      c("political_authority", "religious_authority"),
      c("political_authority", "religious_authority")
    )
  )
stan_code <- coev_make_stancode(
  data = authority$data,
  variables = list(
    political_authority = "ordered_logistic",
    religious_authority = "ordered_logistic"
  ),
  id = "language",
  tree = authority$phylogeny,
  effects_mat = effects_mat
)

# include longitude/latitude values
stan_code <- coev_make_stancode(
  data = authority$data,
  variables = list(
    political_authority = "ordered_logistic",
    religious_authority = "ordered_logistic"
  ),
  id = "language",
  tree = authority$phylogeny,
  lon_lat = authority$coordinates
)

# include measurement error
d <- authority$data
d$x <- rnorm(nrow(d))
d$y <- rnorm(nrow(d))
d$x_std_err <- rexp(nrow(d))
d$y_std_err <- rexp(nrow(d))
stan_code <- coev_make_stancode(
  data = d,
  variables = list(
    x = "normal",
    y = "normal"
  ),
  id = "language",
  tree = authority$phylogeny,
  measurement_error = list(
    x = "x_std_err",
    y = "y_std_err"
  )
)

# set manual priors
stan_code <- coev_make_stancode(
  data = authority$data,
  variables = list(
    political_authority = "ordered_logistic",
    religious_authority = "ordered_logistic"
  ),
  id = "language",
  tree = authority$phylogeny,
  prior = list(A_offdiag = "normal(0, 2)")
)

}
\references{
Ringen, E., Martin, J. S., & Jaeggi, A. (2021). Novel phylogenetic methods
reveal that resource-use intensification drives the evolution of "complex"
societies. \emph{EcoEvoRXiv}. \code{doi:10.32942/osf.io/wfp95}

Sheehan, O., Watts, J., Gray, R. D., Bulbulia, J., Claessens, S., Ringen,
E. J., & Atkinson, Q. D. (2023). Coevolution of religious and political
authority in Austronesian societies. \emph{Nature Human Behaviour},
\emph{7}(1), 38-45. \code{10.1038/s41562-022-01471-y}
}
\seealso{
\code{\link{coev_make_standata}}, \code{\link{coev_fit}}

Other fit:
\code{\link[=coev_fit]{coev_fit()}},
\code{\link[=coev_make_standata]{coev_make_standata()}}
}
\author{
Scott Claessens \email{scott.claessens@gmail.com}, Erik Ringen
\email{erikjacob.ringen@uzh.ch}
}
\concept{fit}
