% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coev_plot_cross.R
\name{coev_plot_cross}
\alias{coev_plot_cross}
\title{Plot cross-selection effects from a fitted \code{coevfit} object}
\usage{
coev_plot_cross(object, ...)
}
\arguments{
\item{object}{An object of class \code{coevfit}}

\item{...}{Additional arguments passed to \code{tidybayes::stat_slabinterval}}
}
\value{
A \code{ggplot} object
}
\description{
Plot cross-selection effects from a fitted \code{coevfit} object
}
\examples{
\dontrun{
# simulate data
n <- 20
tree <- ape::rcoal(n)
d <- data.frame(
  id = tree$tip.label,
  x = rbinom(n, size = 1, prob = 0.5),
  y = ordered(sample(1:4, size = n, replace = TRUE))
)
# fit dynamic coevolutionary model
m <- coev_fit(
  data = d,
  variables = list(
    x = "bernoulli_logit",
    y = "ordered_logistic"
  ),
  id = "id",
  tree = tree,
  # additional arguments for cmdstanr::sample()
  chains = 4,
  parallel_chains = 4,
  iter_warmup = 500,
  seed = 1
)
# plot cross selection effects
coev_plot_cross(m)
}
}
