% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coev_plot_pred_series.R
\name{coev_plot_pred_series}
\alias{coev_plot_pred_series}
\title{Plot a predicted coevolutionary time series from a fitted \code{coevfit}
object}
\usage{
coev_plot_pred_series(object, prob = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{coevfit}}

\item{prob}{A value between 0 and 1 indicating the desired probability
to be covered by the uncertainty intervals. The default is 0.95.}

\item{...}{Additional arguments passed to \code{\link{coev_pred_series}}}
}
\value{
A ggplot object
}
\description{
This function plots a predicted coevolutionary time series using the
estimated parameters from a fitted \code{coevfit} model. By default, the
plot uses the posterior ancestral states estimated by the model as the
starting values, but users can also set their own starting values for
traits. Plots can be generated with or without stochastic drift. For more
details on the underlying predictive function, see
\code{help(coev_pred_series)}.
}
\examples{
\dontrun{
# fit dynamic coevolutionary model
fit <- coev_fit(
  data = authority$data,
  variables = list(
    political_authority = "ordered_logistic",
    religious_authority = "ordered_logistic"
  ),
  id = "language",
  tree = authority$phylogeny,
  # additional arguments for cmdstanr::sample()
  chains = 4,
  parallel_chains = 4,
  seed = 1
  )

# simulated trait co-evolution
coev_plot_pred_series(
  object = fit,
  stochastic = TRUE
  )

# expected trait co-evolution, no drift
coev_plot_pred_series(
  object = fit,
  stochastic = FALSE,
  eta_anc = list(political_authority = -2, religious_authority = 1.5)
  )
}

}
\seealso{
\code{\link{coev_pred_series}}

Other plot:
\code{\link[=coev_plot_delta_theta]{coev_plot_delta_theta()}},
\code{\link[=coev_plot_flowfield]{coev_plot_flowfield()}},
\code{\link[=coev_plot_predictive_check]{coev_plot_predictive_check()}},
\code{\link[=coev_plot_selection_gradient]{coev_plot_selection_gradient()}},
\code{\link[=coev_plot_trait_values]{coev_plot_trait_values()}},
\code{\link[=plot.coevfit]{plot.coevfit()}}
}
\author{
Scott Claessens \email{scott.claessens@gmail.com},
Erik Ringen \email{erikjacob.ringen@uzh.ch}
}
\concept{plot}
