% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coev_plot_predictive_check.R
\name{coev_plot_predictive_check}
\alias{coev_plot_predictive_check}
\title{Plot predictive check from a fitted \code{coevfit} object}
\usage{
coev_plot_predictive_check(
  object,
  variables = NULL,
  ndraws = NULL,
  tree_id = NULL
)
}
\arguments{
\item{object}{An object of class \code{coevfit}}

\item{variables}{If NULL (default), the function returns a list of plots for
all coevolving variables from the model. Otherwise, a character vector
declaring the variables to be included in the list of plots.}

\item{ndraws}{An integer indicating the number of draws to return. The
default and maximum number of draws is the size of the posterior sample.}

\item{tree_id}{An integer indicating the tree ID to use when making
posterior predictions. Set to \code{NULL} by default, which will use draws
from every tree, integrating phylogenetic uncertainty.}
}
\value{
A list of \code{ggplot} objects
}
\description{
Plot posterior (or prior) predictive checks for variables from a fitted
\code{coevfit} object, with help from the \pkg{bayesplot} package.
}
\examples{
\dontrun{
# fit dynamic coevolutionary model
fit <- coev_fit(
  data = authority$data,
  variables = list(
    political_authority = "ordered_logistic",
    religious_authority = "ordered_logistic"
  ),
  id = "language",
  tree = authority$phylogeny,
  # additional arguments for cmdstanr::sample()
  chains = 4,
  parallel_chains = 4,
  seed = 1
  )

# plot predictive checks for all variables
coev_plot_predictive_check(fit)
}

}
\seealso{
Other plot:
\code{\link[=coev_plot_delta_theta]{coev_plot_delta_theta()}},
\code{\link[=coev_plot_flowfield]{coev_plot_flowfield()}},
\code{\link[=coev_plot_pred_series]{coev_plot_pred_series()}},
\code{\link[=coev_plot_selection_gradient]{coev_plot_selection_gradient()}},
\code{\link[=coev_plot_trait_values]{coev_plot_trait_values()}},
\code{\link[=plot.coevfit]{plot.coevfit()}}
}
\author{
Scott Claessens \email{scott.claessens@gmail.com}, Erik Ringen
\email{erikjacob.ringen@uzh.ch}
}
\concept{plot}
