% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coev_plot_selection_matrix.R
\name{coev_plot_selection_matrix}
\alias{coev_plot_selection_matrix}
\title{Plot selection matrix from a fitted \code{coevfit} object}
\usage{
coev_plot_selection_matrix(
  object,
  variables = NULL,
  std = FALSE,
  prob = 0.66,
  prob_outer = 0.95
)
}
\arguments{
\item{object}{An object of class \code{coevfit}}

\item{variables}{If NULL (default), the plot includes all coevolving
variables from the model. Otherwise, a character vector of length >= 2
declaring the variables to be included in the plot.}

\item{std}{Logical. If \code{FALSE} (default), values are not standardised.
If \code{TRUE}, values are scaled by the standard deviation of each
variable among taxa.}

\item{prob}{Probability mass to include in the inner interval. Default is
0.66 (66\% interval).}

\item{prob_outer}{Probability mass to include in the outer interval. Default
is 0.95 (95\% interval).}
}
\value{
A \code{ggplot} object
}
\description{
Plot the selection matrix (i.e., the A matrix) for all trait pairs from a
fitted \code{coevfit} object. Values can be standardised for easier
interpretation.
}
\examples{
\dontrun{
# fit dynamic coevolutionary model
fit <- coev_fit(
  data = authority$data,
  variables = list(
    political_authority = "ordered_logistic",
    religious_authority = "ordered_logistic"
  ),
  id = "language",
  tree = authority$phylogeny,
  # additional arguments for cmdstanr::sample()
  chains = 4,
  parallel_chains = 4,
  seed = 1
  )

# plot selection matrix
coev_plot_selection_matrix(fit)

# standardise the cross-selection effects
coev_plot_selection_matrix(fit, std = TRUE)
}

}
\references{
Ringen, E., Martin, J. S., & Jaeggi, A. (2021). Novel phylogenetic methods
reveal that resource-use intensification drives the evolution of "complex"
societies. \emph{EcoEvoRXiv}. \code{doi:10.32942/osf.io/wfp95}

Sheehan, O., Watts, J., Gray, R. D., Bulbulia, J., Claessens, S., Ringen,
E. J., & Atkinson, Q. D. (2023). Coevolution of religious and political
authority in Austronesian societies. \emph{Nature Human Behaviour},
\emph{7}(1), 38-45. \code{10.1038/s41562-022-01471-y}
}
\author{
Scott Claessens \email{scott.claessens@gmail.com}, Erik Ringen
\email{erikjacob.ringen@uzh.ch}
}
