% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutpie_wrapper.R
\name{draws}
\alias{draws}
\title{Generic method for extracting draws}
\usage{
draws(x, ...)
}
\arguments{
\item{x}{A fit object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A draws_array object.
}
\description{
Generic function for extracting draws from fit objects.
Currently supports nutpie_fit objects.
}
\keyword{internal}
