% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pymc_wrapper.R
\name{metadata.pymc_fit}
\alias{metadata.pymc_fit}
\title{Extract metadata from pymc_fit object}
\usage{
\method{metadata}{pymc_fit}(x)
}
\arguments{
\item{x}{A pymc_fit object.}
}
\value{
A list containing metadata.
}
\description{
Extract metadata from pymc_fit object
}
