% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutpie_wrapper.R
\name{metadata}
\alias{metadata}
\title{Generic method for extracting metadata}
\usage{
metadata(x)
}
\arguments{
\item{x}{A fit object.}
}
\value{
A list containing metadata.
}
\description{
Generic function for extracting metadata from fit objects.
Currently supports nutpie_fit objects.
}
