% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutpie_helpers.R
\name{nutpie_compile_stan_model}
\alias{nutpie_compile_stan_model}
\title{Compile a Stan model using nutpie}
\usage{
nutpie_compile_stan_model(stan_code)
}
\arguments{
\item{stan_code}{Character string containing Stan model code.}
}
\value{
Compiled Stan model object (Python object via reticulate).
}
\description{
Compiles a Stan model string using nutpie's compile_stan_model
function. This is a wrapper around nutpie's Python API via reticulate.
}
\examples{
\dontrun{
stan_code <- "
data {
  int<lower=0> N;
  vector[N] y;
}
parameters {
  real mu;
}
model {
  mu ~ normal(0, 1);
  y ~ normal(mu, 1);
}
"
compiled <- nutpie_compile_stan_model(stan_code)
}

}
