% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{pairs.coevfit}
\alias{pairs.coevfit}
\title{Create a matrix of output plots from a \code{coevfit} object}
\usage{
\method{pairs}{coevfit}(x, parameters = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{coevfit}}

\item{parameters}{If NULL (default), the function returns plots for the main
parameters from the model (the selection matrix "A", the drift matrix "Q",
and the continuous time intercepts "b"). Otherwise, a character vector
declaring the parameters to plot (strictly case-sensitive).}

\item{...}{Further arguments to be passed to
\code{\link[bayesplot:MCMC-scatterplots]{mcmc_pairs}}.}
}
\value{
A matrix of output plots
}
\description{
A \code{\link[graphics:pairs]{pairs}} method that is customized for MCMC
output.
}
\details{
For a detailed description see
\code{\link[bayesplot:MCMC-scatterplots]{mcmc_pairs}}.
}
\examples{
\dontrun{
# fit dynamic coevolutionary model
fit <- coev_fit(
  data = authority$data,
  variables = list(
    political_authority = "ordered_logistic",
    religious_authority = "ordered_logistic"
  ),
  id = "language",
  tree = authority$phylogeny,
  # additional arguments for cmdstanr::sample()
  chains = 4,
  parallel_chains = 4,
  seed = 1
  )

# print pairs plot
pairs(fit)
}

}
\seealso{
\code{\link{plot.coevfit}}
}
\author{
Scott Claessens \email{scott.claessens@gmail.com}
}
