% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standata.R
\name{standata}
\alias{standata}
\title{Expose the Stan data list for a fitted model represented by a \code{coevfit}
object}
\usage{
standata(object)
}
\arguments{
\item{object}{An object of class \code{coevfit}.}
}
\value{
Named list of data for Stan
}
\description{
Expose the Stan data list for a fitted model represented by a \code{coevfit}
object
}
\examples{
\dontrun{
# fit dynamic coevolutionary model
fit <- coev_fit(
  data = authority$data,
  variables = list(
    political_authority = "ordered_logistic",
    religious_authority = "ordered_logistic"
  ),
  id = "language",
  tree = authority$phylogeny,
  # additional arguments for cmdstanr::sample()
  chains = 4,
  parallel_chains = 4,
  seed = 1
  )

# expose stan data
standata(fit)
}

}
\concept{expose}
