% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coev_calculate_delta_theta.R
\name{coev_calculate_delta_theta}
\alias{coev_calculate_delta_theta}
\title{Calculate delta theta from a fitted \code{coevfit} object}
\usage{
coev_calculate_delta_theta(object, response, predictor)
}
\arguments{
\item{object}{An object of class \code{coevfit}}

\item{response}{A character string equal to one of the coevolving variables
in the model}

\item{predictor}{A character string equal to one of the coevolving variables
in the model}
}
\value{
Posterior samples in the draws_array format
}
\description{
Calculate delta theta from a fitted \code{coevfit} object
}
\examples{
\dontrun{
# fit dynamic coevolutionary model
fit <- coev_fit(
  data = authority$data,
  variables = list(
    political_authority = "ordered_logistic",
    religious_authority = "ordered_logistic"
  ),
  id = "language",
  tree = authority$phylogeny,
  # additional arguments for cmdstanr::sample()
  chains = 4,
  parallel_chains = 4,
  seed = 1
  )

# calculate delta theta
coev_calculate_delta_theta(
  object = fit,
  response = "political_authority",
  predictor = "religious_authority"
  )
}
}
