% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coev_calculate_delta_theta.R
\name{coev_calculate_delta_theta}
\alias{coev_calculate_delta_theta}
\title{Calculate delta theta from a fitted \code{coevfit} object}
\usage{
coev_calculate_delta_theta(object, response, predictor)
}
\arguments{
\item{object}{An object of class \code{coevfit}}

\item{response}{A character string equal to one of the coevolving variables
in the model}

\item{predictor}{A character string equal to one of the coevolving variables
in the model}
}
\value{
Posterior samples in the draws_array format
}
\description{
Calculate delta theta from a fitted \code{coevfit} object
}
\examples{
\dontrun{
# simulate data
n <- 20
tree <- ape::rcoal(n)
d <- data.frame(
  id = tree$tip.label,
  x = rbinom(n, size = 1, prob = 0.5),
  y = ordered(sample(1:4, size = n, replace = TRUE))
)
# fit dynamic coevolutionary model
m <- coev_fit(
  data = d,
  variables = list(
    x = "bernoulli_logit",
    y = "ordered_logistic"
  ),
  id = "id",
  tree = tree,
  # additional arguments for cmdstanr::sample()
  chains = 4,
  parallel_chains = 4,
  iter_warmup = 500,
  seed = 1
)
# get delta theta
coev_calculate_delta_theta(m, response = "y", predictor = "x")
}
}
