% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coev_fit.R
\name{coev_fit}
\alias{coev_fit}
\title{Fit Bayesian dynamic coevolutionary model in Stan}
\usage{
coev_fit(data, variables, id, tree, prior = NULL, ...)
}
\arguments{
\item{data}{An object of class \code{data.frame} (or one that can be coerced
to that class) containing data of all variables used in the model.}

\item{variables}{A named list identifying variables that should coevolve in
the model and their associated response distributions (e.g.
\code{list(var1 = bernoulli_logit, var2 = ordered_logistic)}). Must identify
at least two variables. Variable names must refer to valid column names in data.
Currently, the only supported response distributions are \code{bernoulli_logit}
and \code{ordered_logistic}.}

\item{id}{A string identifying the variable in the data that links rows to tips
on the phylogeny. Must refer to a valid column name in the data. The id column
must exactly match the tip labels in the phylogeny.}

\item{tree}{A phylogenetic tree object of class \code{phylo}.}

\item{prior}{A list of priors for the model.}

\item{...}{Additional arguments for \pkg{Stan}.}
}
\value{
The model.
}
\description{
Fit Bayesian dynamic coevolutionary model in Stan
}
\examples{
#coev_fit()
}
