% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coev_fit.R
\name{coev_fit}
\alias{coev_fit}
\title{Fit Bayesian dynamic coevolutionary model in Stan}
\usage{
coev_fit(
  data,
  variables,
  id,
  tree,
  effects_mat = NULL,
  complete_cases = FALSE,
  dist_mat = NULL,
  dist_cov = "exp_quad",
  prior = NULL,
  scale = TRUE,
  estimate_Q_offdiag = TRUE,
  log_lik = FALSE,
  prior_only = FALSE,
  adapt_delta = 0.95,
  ...
)
}
\arguments{
\item{data}{An object of class \code{data.frame} (or one that can be coerced
to that class) containing data of all variables used in the model.}

\item{variables}{A named list identifying variables that should coevolve in
the model and their associated response distributions as character strings
(e.g. \code{list(var1 = "bernoulli_logit", var2 = "ordered_logistic")}).
Must identify at least two variables. Variable names must refer to valid
column names in data. Currently, the only supported response distributions
are \code{bernoulli_logit}, \code{ordered_logistic},
\code{poisson_softplus}, \code{negative_binomial_softplus}, \code{normal},
and \code{gamma_log}.}

\item{id}{A character of length one identifying the variable in the data that
links rows to tips on the phylogeny. Must refer to a valid column name in
the data. The id column must exactly match the tip labels in the phylogeny.}

\item{tree}{A phylogenetic tree object of class \code{phylo} or
\code{multiPhylo}. The tree(s) must be rooted and must include positive
non-zero branch lengths. All trees in \code{multiPhylo} objects must have
the same number of internal nodes and branches.}

\item{effects_mat}{(optional) A boolean matrix with row and column names
exactly matching the variables declared for the model. If not specified,
all cross-lagged effects will be estimated in the model. If specified, the
model will only estimate cross-lagged effects where cells in the matrix are
TRUE and will ignore cross-lagged effects where cells in the matrix are
FALSE. In the matrix, columns represent predictor variables and rows
represent outcome variables. All autoregressive effects (e.g., X -> X) must
be TRUE in the matrix.}

\item{complete_cases}{(optional) Logical. If \code{FALSE} (default), all
missing values are imputed by the model. If \code{TRUE}, taxa with missing
data are excluded.}

\item{dist_mat}{(optional) A distance matrix with row and column names
exactly matching the tip labels in the phylogeny. If specified, the model
will additionally control for spatial location by including a separate
Gaussian Process over locations for every coevolving variable in the model.}

\item{dist_cov}{A string specifying the covariance kernel used for Gaussian
Processes over locations. Currently supported are \code{"exp_quad"}
(exponentiated-quadratic kernel; default), \code{"exponential"}
(exponential kernel), and \code{"matern32"} (Matern 3/2 kernel).}

\item{prior}{(optional) A named list of priors for the model. If not
specified, the model uses default priors (see \code{help(coev_fit)}).
Alternatively, the user can specify a named list of priors. The list must
contain non-duplicated entries for any of the following parameters: the
autoregressive effects (\code{A_diag}), the cross effects
(\code{A_offdiag}), the Cholesky factor for the drift matrix (\code{L_R}),
the drift std. dev. parameters (\code{Q_sigma}), the continuous time
intercepts (\code{b}), the ancestral states for the traits
(\code{eta_anc}), the cutpoints for ordinal variables (\code{c}), the
overdispersion parameters for negative binomial variables (\code{phi}),
the shape parameters for gamma variables (\code{shape}), the sigma
parameters for Gaussian Processes over locations (\code{sigma_dist}), the
rho parameters for Gaussian Processes over locations (\code{rho_dist}), the
standard deviation parameters for non-phylogenetic group-level varying
effects (\code{sigma_group}), and the Cholesky factor for the
non-phylogenetic group-level correlation matrix (\code{L_group}). These
must be entered with valid prior strings, e.g.
\code{list(A_offdiag = "normal(0, 2)")}. Invalid prior strings will throw
an error when the function internally checks the syntax of resulting Stan
code.}

\item{scale}{Logical. If \code{TRUE} (default), variables following the
\code{normal} and \code{gamma_log} response distributions are scaled before
fitting the model. Continuous variables following the \code{normal}
distribution are standardised (e.g., mean centered and divided by their
standard deviation) and positive real variables following the
\code{gamma_log} distribution are divided by the mean value without
centering. This approach is recommended when using default priors to
improve efficiency and ensure accurate inferences. If \code{FALSE},
variables are left unscaled for model fitting. In this case, users should
take care to set sensible priors on variables.}

\item{estimate_Q_offdiag}{Logical. If \code{TRUE} (default), the model
estimates the off-diagonals for the \eqn{Q} drift matrix (i.e., correlated
drift). If \code{FALSE}, the off-diagonals for the \eqn{Q} drift matrix
are set to zero.}

\item{log_lik}{Logical. Set to \code{FALSE} by default. If \code{TRUE}, the
model returns the pointwise log likelihood, which can be used to calculate
WAIC and LOO.}

\item{prior_only}{Logical. If \code{FALSE} (default), the model is fitted to
the data and returns a posterior distribution. If \code{TRUE}, the model
samples from the prior only, ignoring the likelihood.}

\item{adapt_delta}{Argument for \pkg{cmdstanr::sample()}. Default is 0.95.}

\item{...}{Additional arguments for \pkg{cmdstanr::sample()}.}
}
\value{
Fitted model of class \code{coevfit}.
}
\description{
Fit Bayesian dynamic coevolutionary models in \pkg{Stan} for full Bayesian
inference. The model allows users to assess causal directionality
(X -> Y vs. Y -> X) and contingencies (X, then Y) in evolution. Several data
types are supported, including binary, ordinal, count, continuous, and
positive real variables. The model can additionally account for missing data,
repeated observations, and controls for spatial proximity. Model fit can be
assessed and compared with posterior predictive checks and cross-validation.
}
\details{
The \code{coev_fit} function generates the \pkg{Stan} code for the model,
generates the data list, and then compiles and fits the model using the
\pkg{cmdstanr} package.

Fit a Bayesian dynamic coevolutionary model in Stan. A general
overview is provided in the vignette \code{vignette("coevolve")}

\bold{Available response distributions}

Variables that should coevolve in the model are declared in the
\code{variables} argument, using a named list with associated response
distributions. For example: \code{list(x = "bernoulli_logit", y =
  "ordered_logistic")}. Currently, the only supported response distributions
are \code{bernoulli_logit}, \code{ordered_logistic},
\code{poisson_softplus}, \code{negative_binomial_softplus}, \code{normal},
and \code{gamma_log}.

\bold{Default prior distributions}

If priors are not explicitly declared by the user, default priors are used.
Default priors were chosen to be weakly regularising, which improves model
fitting and conservatism in parameter estimates. Default priors for Stan
parameters are as follows:
\itemize{
\item \code{A_diag} (autoregressive effects) = \code{std_normal()}
\item \code{A_offdiag} (cross effects) = \code{std_normal()}
\item \code{L_R} (Cholesky factor for drift matrix) =
\code{lkj_corr_cholesky(4)}
\item \code{Q_sigma} (drift std. dev. parameters) = \code{std_normal()}
\item \code{b} (continuous time intercepts) = \code{std_normal()}
\item \code{eta_anc} (trait ancestral states) = \code{std_normal()}
\item \code{c} (ordinal cutpoints) = \code{normal(0, 2)}
\item \code{shape} (shape parameters for gamma distributions) =
\code{gamma(0.01, 0.01)}
\item \code{sigma_dist} (sigma for Gaussian process over locations) =
\code{exponential(1)}
\item \code{rho_dist} (rho for Gaussian process over locations) =
\code{exponential(5)}
\item \code{sigma_group} (standard deviation for group-level varying effects) =
\code{exponential(1)}
\item \code{L_group} (Cholesky factor for group-level varying effects) =
\code{lkj_corr_cholesky(2)}
}

The default prior for \code{phi} (the overdispersion parameter for the
negative-binomial distribution) is scaled automatically based on the
variance of the data.

We recommend that users assess the suitability of these default priors by
fitting the model with \code{prior_only = TRUE} and then plotting prior
predictive checks for all variables using the
\code{coev_plot_predictive_check()} function.

\bold{Handling missing data}

In order to retain the most information, the \code{coev_fit()} function
automatically imputes all missing values. To turn off this behaviour and
exclude taxa with missing data, set \code{complete_cases = FALSE}.

\bold{Dealing with repeated observations}

If taxa appear in the dataset multiple times (i.e., there are repeated
observations), the model will automatically include non-phylogenetic
group-level varying effects and correlations to account for this
clustering. These parameters represent the between-taxa variation and
correlation that remains after accounting for the coevolutionary process.

\bold{Controlling for spatial location}

If users declare a distance matrix with the \code{dist_mat} argument,
the model will include several Gaussian processes (one per coevolving
variable) over spatial locations.
}
\examples{
\dontrun{
# fit dynamic coevolutionary model
fit <- coev_fit(
  data = authority$data,
  variables = list(
    political_authority = "ordered_logistic",
    religious_authority = "ordered_logistic"
  ),
  id = "language",
  tree = authority$phylogeny,
  # additional arguments for cmdstanr::sample()
  chains = 4,
  parallel_chains = 4,
  seed = 1
  )

# print model summary
summary(fit)
}

}
\references{
Ringen, E., Martin, J. S., & Jaeggi, A. (2021). Novel phylogenetic methods
reveal that resource-use intensification drives the evolution of "complex"
societies. \emph{EcoEvoRXiv}. \code{doi:10.32942/osf.io/wfp95}

Sheehan, O., Watts, J., Gray, R. D., Bulbulia, J., Claessens, S., Ringen,
E. J., & Atkinson, Q. D. (2023). Coevolution of religious and political
authority in Austronesian societies. \emph{Nature Human Behaviour},
\emph{7}(1), 38-45. \code{10.1038/s41562-022-01471-y}
}
\seealso{
\code{\link{coev_make_stancode}}, \code{\link{coev_make_standata}}
}
\author{
Scott Claessens \email{scott.claessens@gmail.com}, Erik Ringen
\email{erikjacob.ringen@uzh.ch}
}
