% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coev_make_stancode.R
\name{coev_make_stancode}
\alias{coev_make_stancode}
\title{Make Stan code for dynamic coevolutionary model}
\usage{
coev_make_stancode(data, variables, id, tree, prior = NULL)
}
\arguments{
\item{data}{An object of class \code{data.frame} (or one that can be coerced
to that class) containing data of all variables used in the model.}

\item{variables}{A named list identifying variables that should coevolve in
the model and their associated response distributions as character strings (e.g.
\code{list(var1 = "bernoulli_logit", var2 = "ordered_logistic")}). Must identify
at least two variables. Variable names must refer to valid column names in data.
Currently, the only supported response distributions are \code{bernoulli_logit}
and \code{ordered_logistic}.}

\item{id}{A string identifying the variable in the data that links rows to tips
on the phylogeny. Must refer to a valid column name in the data. The id column
must exactly match the tip labels in the phylogeny.}

\item{tree}{A phylogenetic tree object of class \code{phylo}.}

\item{prior}{A list of priors for the model.}
}
\value{
A character string containing the \pkg{Stan} code to fit the dynamic coevolutionary model.
}
\description{
Make Stan code for dynamic coevolutionary model
}
\examples{
# simulate data
set.seed(1)
n <- 20
tree <- ape::rtree(n)
d <- data.frame(
   id = tree$tip.label,
   x = rbinom(n, size = 1, prob = 0.5),
   y = ordered(sample(1:4, size = n, replace = TRUE))
)
# make stan code
coev_make_stancode(
   data = d,
   variables = list(
       x = "bernoulli_logit",
       y = "ordered_logistic"
   ),
   id = "id",
   tree = tree
)
}
