% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coev_plot_delta_theta.R
\name{coev_plot_delta_theta}
\alias{coev_plot_delta_theta}
\title{Plot delta theta values from a fitted \code{coevfit} object}
\usage{
coev_plot_delta_theta(object, variables = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{coevfit}}

\item{variables}{If NULL (default), the plot includes all coevolving
variables from the model. Otherwise, a character vector of length >= 2
declaring the variables to be included in the plot.}

\item{...}{Additional arguments passed to \code{ggdist::stat_slabinterval}}
}
\value{
A \code{ggplot} object
}
\description{
Plot delta theta values from a fitted \code{coevfit} object
}
\examples{
\dontrun{
# simulate data
n <- 20
tree <- ape::rcoal(n)
d <- data.frame(
  id = tree$tip.label,
  x = rbinom(n, size = 1, prob = 0.5),
  y = ordered(sample(1:4, size = n, replace = TRUE))
)
# fit dynamic coevolutionary model
m <- coev_fit(
  data = d,
  variables = list(
    x = "bernoulli_logit",
    y = "ordered_logistic"
  ),
  id = "id",
  tree = tree,
  # additional arguments for cmdstanr::sample()
  chains = 4,
  parallel_chains = 4,
  iter_warmup = 500,
  seed = 1
)
# plot delta theta values for all effects
coev_plot_delta_theta(m)
}
}
