% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coev_plot_pred_series.R
\name{coev_plot_pred_series}
\alias{coev_plot_pred_series}
\title{Plot a predicted co-evolutionary time series from a fitted \code{coevfit}
object}
\usage{
coev_plot_pred_series(object, prob = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{coevfit}}

\item{prob}{A value between 0 and 1 indicating the desired probability
to be covered by the uncertainty intervals. The default is 0.95.}

\item{...}{Additional arguments passed to \code{\link{coev_pred_series}}}
}
\value{
A ggplot object.
}
\description{
Plot a predicted co-evolutionary time series from a fitted \code{coevfit}
object
}
\examples{
\dontrun{
# fit dynamic coevolutionary model
fit <- coev_fit(
  data = authority$data,
  variables = list(
    political_authority = "ordered_logistic",
    religious_authority = "ordered_logistic"
  ),
  id = "language",
  tree = authority$phylogeny,
  # additional arguments for cmdstanr::sample()
  chains = 4,
  parallel_chains = 4,
  seed = 1
  )

# simulated trait co-evolution
coev_plot_pred_series(
  object = fit,
  stochastic = T
  )

# expected trait co-evolution, no drift
coev_plot_pred_series(
  object = fit,
  stochastic = F
  )
}

}
\author{
Scott Claessens \email{scott.claessens@gmail.com}, Erik Ringen
\email{erikjacob.ringen@uzh.ch}
}
