% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coev_plot_predictive_check.R
\name{coev_plot_predictive_check}
\alias{coev_plot_predictive_check}
\title{Plot posterior predictive check from a fitted \code{coevfit} object}
\usage{
coev_plot_predictive_check(object, variables = NULL, ndraws = NULL)
}
\arguments{
\item{object}{An object of class \code{coevfit}}

\item{variables}{If NULL (default), the function returns a list of plots for
all coevolving variables from the model. Otherwise, a character vector
declaring the variables to be included in the list of plots.}

\item{ndraws}{An integer indicating the number of draws to return. The
default and maximum number of draws is the size of the posterior sample.}
}
\value{
A list of \code{ggplot} objects
}
\description{
Plot posterior predictive check from a fitted \code{coevfit} object
}
\examples{
\dontrun{
# simulate data
n <- 20
tree <- ape::rcoal(n)
d <- data.frame(
  id = tree$tip.label,
  x = rbinom(n, size = 1, prob = 0.5),
  y = ordered(sample(1:4, size = n, replace = TRUE))
)
# fit dynamic coevolutionary model
m <- coev_fit(
  data = d,
  variables = list(
    x = "bernoulli_logit",
    y = "ordered_logistic"
  ),
  id = "id",
  tree = tree,
  # additional arguments for cmdstanr::sample()
  chains = 4,
  parallel_chains = 4,
  iter_warmup = 500,
  seed = 1
)
# plot predictive checks for all variables
coev_plot_predictive_check(m)
}
}
