% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coev_plot_selection_gradient.R
\name{coev_plot_selection_gradient}
\alias{coev_plot_selection_gradient}
\title{Plot selection gradient heatmap from a fitted \code{coevfit} object}
\usage{
coev_plot_selection_gradient(object, var1, var2, contour = FALSE)
}
\arguments{
\item{object}{An object of class \code{coevfit}}

\item{var1}{A character string equal to one of the coevolving variables in
the model}

\item{var2}{A character string equal to one of the coevolving variables in
the model}

\item{contour}{Logical (defaults to FALSE); whether to show white contour
lines to indicate where selection is stronger than drift}
}
\value{
A \code{ggplot} object
}
\description{
Plot selection gradient heatmap from a fitted \code{coevfit} object
}
\examples{
\dontrun{
# fit dynamic coevolutionary model
fit <- coev_fit(
  data = authority$data,
  variables = list(
    political_authority = "ordered_logistic",
    religious_authority = "ordered_logistic"
  ),
  id = "language",
  tree = authority$phylogeny,
  # additional arguments for cmdstanr::sample()
  chains = 4,
  parallel_chains = 4,
  seed = 1
  )

# plot selection gradient
coev_plot_selection_gradient(
  object = fit,
  var1 = "political_authority",
  var2 = "religious_authority"
  )
}
}
