% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coev_plot_selection_gradient.R
\name{coev_plot_selection_gradient}
\alias{coev_plot_selection_gradient}
\title{Plot selection gradient heatmap from a fitted \code{coevfit} object}
\usage{
coev_plot_selection_gradient(object, var1, var2, contour = FALSE)
}
\arguments{
\item{object}{An object of class \code{coevfit}}

\item{var1}{A character string equal to one of the coevolving variables in the model}

\item{var2}{A character string equal to one of the coevolving variables in the model}

\item{countour}{Logical (defaults to FALSE); whether to show white contour lines
to indicate where selection is stronger than drift}
}
\value{
A \code{ggplot} object
}
\description{
Plot selection gradient heatmap from a fitted \code{coevfit} object
}
\examples{
\dontrun{
# simulate data
n <- 20
tree <- ape::rcoal(n)
d <- data.frame(
  id = tree$tip.label,
  x = rbinom(n, size = 1, prob = 0.5),
  y = ordered(sample(1:4, size = n, replace = TRUE))
)
# fit dynamic coevolutionary model
m <- coev_fit(
  data = d,
  variables = list(
    x = "bernoulli_logit",
    y = "ordered_logistic"
  ),
  id = "id",
  tree = tree,
  # additional arguments for cmdstanr::sample()
  chains = 4,
  parallel_chains = 4,
  iter_warmup = 500,
  seed = 1
)
# plot selection gradient
coev_plot_selection_gradient(m)
}
}
