% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coev_plot_selection_gradient.R
\name{coev_plot_selection_gradient}
\alias{coev_plot_selection_gradient}
\title{Plot selection gradient heatmap from a fitted \code{coevfit} object}
\usage{
coev_plot_selection_gradient(
  object,
  var1,
  var2,
  contour = FALSE,
  limits = c(-2.5, 2.5)
)
}
\arguments{
\item{object}{An object of class \code{coevfit}}

\item{var1}{A character string equal to one of the coevolving variables in
the model}

\item{var2}{A character string equal to one of the coevolving variables in
the model}

\item{contour}{Logical (defaults to FALSE); whether to show white contour
lines to indicate where selection is stronger than drift}

\item{limits}{A numeric vector of length 2 (defaults to \code{c(-2.5, 2.5)});
specifying the lower limit and the upper limit of the x and y axes.}
}
\value{
A \code{ggplot} object
}
\description{
Plot a heatmap of the selection gradient for two variables from a fitted
\code{coevfit} object.
}
\details{
The selection gradient is operationalised as the ratio of the
change in a trait due to deterministic selection \eqn{\Delta\alpha} to
the change in a trait due to stochastic drift \eqn{\sigma}. Values
between -1 and 1 indicate parameter space where the change due to
drift is greater than change due to selection on the trait. Conversely,
values greater than 1 (or less than -1) indicate parameter space where
positive (or negative) selection is stronger than drift. If three or more
traits were included in the model, other traits are held at their median
values during these computations.

If the plot does not look right, the user might try zooming out from the
default parameter space by setting wider limits. For some variables (e.g.,
continuous and count variables), the default limits may not be suitable.
}
\examples{
\dontrun{
# fit dynamic coevolutionary model
fit <- coev_fit(
  data = authority$data,
  variables = list(
    political_authority = "ordered_logistic",
    religious_authority = "ordered_logistic"
  ),
  id = "language",
  tree = authority$phylogeny,
  # additional arguments for cmdstanr::sample()
  chains = 4,
  parallel_chains = 4,
  seed = 1
  )

# plot selection gradient
coev_plot_selection_gradient(
  object = fit,
  var1 = "political_authority",
  var2 = "religious_authority"
  )
}

}
\references{
Ringen, E., Martin, J. S., & Jaeggi, A. (2021). Novel phylogenetic methods
reveal that resource-use intensification drives the evolution of "complex"
societies. \emph{EcoEvoRXiv}. \code{doi:10.32942/osf.io/wfp95}

Sheehan, O., Watts, J., Gray, R. D., Bulbulia, J., Claessens, S., Ringen,
E. J., & Atkinson, Q. D. (2023). Coevolution of religious and political
authority in Austronesian societies. \emph{Nature Human Behaviour},
\emph{7}(1), 38-45. \code{10.1038/s41562-022-01471-y}
}
\seealso{
\code{\link{coev_plot_flowfield}}
}
\author{
Scott Claessens \email{scott.claessens@gmail.com}, Erik Ringen
\email{erikjacob.ringen@uzh.ch}
}
