% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coev_pred_series.R
\name{coev_pred_series}
\alias{coev_pred_series}
\title{Predict a co-evolutionary time series from a fitted \code{coevfit}
object}
\usage{
coev_pred_series(
  object,
  eta_anc = NULL,
  tmax = 1,
  ntimes = 30,
  ndraws = NULL,
  stochastic = FALSE
)
}
\arguments{
\item{object}{An object of class \code{coevfit}}

\item{eta_anc}{If \code{NULL} default, the starting values for the latent states \eqn{\eta} will be set to the estimated ancestral states at the root from the the \code{coevfit}. Otherwise, a numeric vector with length equal to the number of variable specifying the initial \eqn{\eta}  values.}

\item{tmax}{A positive number indicating the total duration of time for which to predict. Set to 1 by default, corresponding to the entire time depth of the original phylogenetic tree(s).}

\item{ntimes}{A positive integer indicating the total number of discrete time steps over which to make predictions. Each step corresponds to a time difference of dt = tmax/ntimes. Set to 30 by default.}

\item{ndraws}{An integer indicating the number of draws to return. The
default and maximum number of draws is the size of the posterior sample.}

\item{stochastic}{Logical (defaults to \code{FALSE}); indicator of whether predictions should include only the expected co-evolutionary change (deterministic selection), or also stochastic drift.}
}
\value{
An \link{ndraw, ntimes, nvariables} array of predicted \eqn{\eta}  values.
}
\description{
Predict a co-evolutionary time series from a fitted \code{coevfit}
object
}
\examples{
\dontrun{
# fit dynamic coevolutionary model
fit <- coev_fit(
  data = authority$data,
  variables = list(
    political_authority = "ordered_logistic",
    religious_authority = "ordered_logistic"
  ),
  id = "language",
  tree = authority$phylogeny,
  # additional arguments for cmdstanr::sample()
  chains = 4,
  parallel_chains = 4,
  seed = 1
  )

# simulate trait co-evolution
sims <- coev_pred_series(
  object = fit,
  stochastic = T
  )

# expected trait co-evolution, no drift
epreds <- coev_pred_series(
  object = fit,
  stochastic = F
  )
}

}
\author{
Scott Claessens \email{scott.claessens@gmail.com}, Erik Ringen
\email{erikjacob.ringen@uzh.ch}
}
