% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coev_simulate_coevolution.R
\name{coev_simulate_coevolution}
\alias{coev_simulate_coevolution}
\title{Simulate the coevolution of multiple variables in discrete time steps}
\usage{
coev_simulate_coevolution(n, variables, selection_matrix, drift, prob_split)
}
\arguments{
\item{n}{Number of data points in the resulting data frame.}

\item{variables}{A character vector of variable names (e.g.,
\code{c("x","y")})}

\item{selection_matrix}{A numeric matrix determining the strength of
selection between variables. The matrix must have a number of rows and
columns equal to the number of variables and its row and column names must
contain all specified variables. Each cell determines the strength of
selection from the column variable to the row variable. For example, the
cell on the "x" column and the "x" row indicates how much previous values
of x influence future values of x (autocorrelation). By contrast, the cell
on the "x" column and the "y" row indicates how much previous values of x
influence future values of y (cross-lagged effect).}

\item{drift}{A named numeric vector specifying the strength of drift for
different variables. Names must include all specified variables.}

\item{prob_split}{A numeric probability of a species split in any given
timestep.}
}
\value{
List with dataset at final timestep (data), full simulation log
(simulation), and pruned phylogenetic tree (tree).
}
\description{
Simulate the coevolution of multiple variables in discrete time steps
}
\examples{
# simulate coevolution of x and y
n <- 100
variables <- c("x","y")
# x -> y but not vice versa
selection_matrix <- matrix(
  c(
    0.95, 0.00,
    0.80, 0.95
  ),
  nrow = 2,
  byrow = TRUE,
  dimnames = list(variables, variables)
)
drift <- c("x" = 0.05, "y" = 0.05)
prob_split <- 0.05
# run simulation
sim <- coev_simulate_coevolution(n, variables, selection_matrix,
                                 drift, prob_split)
}
