% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coevolve-package.R
\docType{package}
\name{coevolve-package}
\alias{coevolve}
\alias{coevolve-package}
\title{coevolve: Fit Bayesian Generalized Dynamic Phylogenetic Models using 'Stan' or 'JAX'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Fit Bayesian generalized dynamic phylogenetic models using 'Stan' or 'JAX'. The package allows an arbitrary number of variables with different response distributions to coevolve on a phylogenetic tree via an Ornstein-Uhlenbeck evolutionary process, allowing users to assess directionality between coevolving variables. The optional JAX backend enables high-performance sampling with NumPyro and potential GPU acceleration via 'reticulate'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ScottClaessens/coevolve}
  \item \url{https://scottclaessens.github.io/coevolve/}
  \item Report bugs at \url{https://github.com/ScottClaessens/coevolve/issues}
}

}
\author{
\strong{Maintainer}: Scott Claessens \email{scott.claessens@gmail.com} (\href{https://orcid.org/0000-0002-3562-6981}{ORCID})

Authors:
\itemize{
  \item Scott Claessens \email{scott.claessens@gmail.com} (\href{https://orcid.org/0000-0002-3562-6981}{ORCID})
  \item Erik Ringen \email{erikjacob.ringen@uzh.ch} (\href{https://orcid.org/0000-0002-3565-6961}{ORCID})
}

Other contributors:
\itemize{
  \item Rob Moss \email{rgmoss@unimelb.edu.au} (\href{https://orcid.org/0000-0002-4568-2012}{ORCID}) [reviewer]
  \item Krzysztof Bartoszek \email{krzbar@protonmail.ch} (\href{https://orcid.org/0000-0002-5816-4345}{ORCID}) [reviewer]
}

}
\keyword{internal}
