% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coevolve-package.R
\docType{package}
\name{coevolve-package}
\alias{coevolve}
\alias{coevolve-package}
\title{coevolve: Fit Bayesian Dynamic Coevolutionary Models using 'Stan'}
\description{
Fit Bayesian dynamic coevolutionary models using 'Stan'. The package allows an abritrary number of variables with different response distributions to coevolve on a phylogenetic tree via an Ornstein-Uhlenbeck evolutionary process, allowing users to assess directionality between coevolving variables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ScottClaessens/coevolve}
}

}
\author{
\strong{Maintainer}: Scott Claessens \email{scott.claessens@gmail.com} (\href{https://orcid.org/0000-0002-3562-6981}{ORCID})

Authors:
\itemize{
  \item Erik Ringen \email{erikjacob.ringen@uzh.ch}
}

}
\keyword{internal}
