% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_samples.R
\name{extract_samples}
\alias{extract_samples}
\title{Extract samples (draws) from a fitted \code{coevfit} object}
\usage{
extract_samples(object)
}
\arguments{
\item{object}{An object of class \code{coevfit}}
}
\value{
Samples in 'rethinking' style list format
}
\description{
Extract samples (draws) from a fitted \code{coevfit} object
}
