% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_samples.R
\name{extract_samples}
\alias{extract_samples}
\title{Extract samples (draws) from a fitted \code{coevfit} object}
\usage{
extract_samples(object)
}
\arguments{
\item{object}{An object of class \code{coevfit}}
}
\value{
Samples in 'rethinking' style list format
}
\description{
Extract samples (draws) from a fitted \code{coevfit} object
}
\examples{
\dontrun{
# fit dynamic coevolutionary model
fit <- coev_fit(
  data = authority$data,
  variables = list(
    political_authority = "ordered_logistic",
    religious_authority = "ordered_logistic"
  ),
  id = "language",
  tree = authority$phylogeny,
  # additional arguments for cmdstanr::sample()
  chains = 4,
  parallel_chains = 4,
  seed = 1
  )

# extract samples from fitted model
extract_samples(fit)
}

}
