% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jax_wrapper.R
\name{metadata.jax_fit}
\alias{metadata.jax_fit}
\title{Extract metadata from jax_fit object}
\usage{
\method{metadata}{jax_fit}(x)
}
\arguments{
\item{x}{A jax_fit object.}
}
\value{
A list containing metadata.
}
\description{
Extract metadata from jax_fit object
}
