% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.coevfit}
\alias{plot.coevfit}
\title{Trace and density plots for \code{coevfit} objects}
\usage{
\method{plot}{coevfit}(
  x,
  parameters = NULL,
  combo = c("dens", "trace"),
  npars = 5,
  plot = TRUE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{coevfit}.}

\item{parameters}{If NULL (default), the function returns a list of plots for
the main parameters from the model (the selection matrix "A", the drift
matrix "Q", and the continuous time intercepts "b"). Otherwise, a character
vector declaring the parameters to plot (strictly case-sensitive).}

\item{combo}{A character vector with at least two elements. Each element of
\code{combo} corresponds to a column in the resulting plot and should be
the name of one of the available
\code{\link[bayesplot:MCMC-overview]{MCMC}} functions (omitting the
\code{mcmc_} prefix).}

\item{npars}{Number of parameters displayed per page (default = 5).}

\item{plot}{Logical. If TRUE (default), plots are plotted in the active
graphic device. If FALSE, plots are not plotted.}

\item{...}{Not used.}
}
\value{
A list of \code{ggplot} objects
}
\description{
Trace and density plots for \code{coevfit} objects
}
\examples{
\dontrun{
# fit dynamic coevolutionary model
fit <- coev_fit(
  data = authority$data,
  variables = list(
    political_authority = "ordered_logistic",
    religious_authority = "ordered_logistic"
  ),
  id = "language",
  tree = authority$phylogeny,
  # additional arguments for cmdstanr::sample()
  chains = 4,
  parallel_chains = 4,
  seed = 1
  )

# print trace and density plots
plot(fit)
}

}
\seealso{
Other plot:
\code{\link[=coev_plot_delta_theta]{coev_plot_delta_theta()}},
\code{\link[=coev_plot_flowfield]{coev_plot_flowfield()}},
\code{\link[=coev_plot_pred_series]{coev_plot_pred_series()}},
\code{\link[=coev_plot_predictive_check]{coev_plot_predictive_check()}},
\code{\link[=coev_plot_selection_gradient]{coev_plot_selection_gradient()}},
\code{\link[=coev_plot_trait_values]{coev_plot_trait_values()}}
}
\concept{plot}
