% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.coevfit}
\alias{print.coevfit}
\alias{print.coevsummary}
\title{Print a summary of a fitted model represented by a \code{coevfit} object}
\usage{
\method{print}{coevfit}(x, digits = 2, ...)
}
\arguments{
\item{x}{An object of class \code{coevfit}}

\item{digits}{The number of significant digits for printing out the summary;
defaults to 2}

\item{...}{Additional arguments that would be passed
to method \code{summary} of \code{coevfit}.}
}
\value{
A printed summary of a \code{coevfit} model object
}
\description{
Print a summary of a fitted model represented by a \code{coevfit} object
}
\examples{
\dontrun{
# fit dynamic coevolutionary model
fit <- coev_fit(
  data = authority$data,
  variables = list(
    political_authority = "ordered_logistic",
    religious_authority = "ordered_logistic"
  ),
  id = "language",
  tree = authority$phylogeny,
  # additional arguments for cmdstanr::sample()
  chains = 4,
  parallel_chains = 4,
  seed = 1
  )

# print summary of model
print(fit)
}

}
\seealso{
\code{\link{summary.coevfit}}
}
