% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.coevfit}
\alias{print.coevfit}
\title{Print the summary of a fitted model represented by a \code{coevfit} object}
\usage{
\method{print}{coevfit}(object, digits = 2)
}
\arguments{
\item{object}{An object of class \code{coevfit}}

\item{digits}{The number of significant digits for printing out the summary; defaults to 2}
}
\description{
Print the summary of a fitted model represented by a \code{coevfit} object
}
