% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_coevfit.R
\name{save_coevfit}
\alias{save_coevfit}
\title{Save a fitted \code{coevfit} object to a file}
\usage{
save_coevfit(object, file, ...)
}
\arguments{
\item{object}{An object of class \code{coevfit}}

\item{file}{A string declaring the path where the file should be saved}

\item{...}{Other arguments to pass to \code{\link[base]{saveRDS}} besides
\code{object} and \code{file}}
}
\value{
An .RDS file containing the fitted \code{coevfit} model object
}
\description{
A wrapper around \code{\link[base]{saveRDS}} that saves a fitted
\code{coevfit} model object while ensuring that all posterior draws and
diagnostics are correctly saved by \pkg{cmdstanr}.
}
