% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stancode.R
\name{stancode}
\alias{stancode}
\title{Expose the stancode of a fitted model represented by a \code{coevfit} object}
\usage{
stancode(object)
}
\arguments{
\item{object}{An object of class \code{coevfit}.}
}
\description{
Expose the stancode of a fitted model represented by a \code{coevfit} object
}
