% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standata.R
\name{standata}
\alias{standata}
\title{Expose the Stan data list for a fitted model represented by a \code{coevfit}
object}
\usage{
standata(object)
}
\arguments{
\item{object}{An object of class \code{coevfit}.}
}
\value{
Named list of data for Stan
}
\description{
Expose the Stan data list for a fitted model represented by a \code{coevfit}
object
}
