% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.coevfit}
\alias{summary.coevfit}
\title{Create a summary of a fitted model represented by a \code{coevfit} object}
\usage{
\method{summary}{coevfit}(object, prob = 0.95, robust = FALSE, ...)
}
\arguments{
\item{object}{An object of class \code{coevfit}.}

\item{prob}{A value between 0 and 1 indicating the desired probability
to be covered by the uncertainty intervals. The default is 0.95.}

\item{robust}{If \code{FALSE} (the default) the mean is used as the measure
of central tendency and the standard deviation as the measure of
variability. If \code{TRUE}, the median and the median absolute deviation
are applied instead.}

\item{...}{Other potential arguments}
}
\description{
Create a summary of a fitted model represented by a \code{coevfit} object
}
\details{
The convergence diagnostics \code{rhat}, \code{ess_bulk}, and
\code{ess_tail} are described in detail in Vehtari et al. (2020).
}
\references{
Aki Vehtari, Andrew Gelman, Daniel Simpson, Bob Carpenter, and
Paul-Christian Bürkner (2020). Rank-normalization, folding, and
localization: An improved R-hat for assessing convergence of
MCMC. \emph{Bayesian Analysis}. 1–28. dpi:10.1214/20-BA1221
}
